+++ crates/rune-cli/src/run/command.rs
@@ -79,32 +79,44 @@ impl Run {
         })?;
 
         let img = self.initialize_image(&rune)?;
+        let mut call = self.load_runtime(&rune, img)?;
 
+        log::info!("The Rune was loaded successfully");
+
+        if cfg!(target_os = "macos") {
+            let ticket = "https://github.com/tensorflow/tensorflow/issues/52300";
+            log::warn!("TensorFlow Lite has a bug where its MacOS CPU backend will occasionally segfault during inference. See {} for more.", ticket);
+        }
+
+        for i in 0..self.repeats {
+            if i > 0 {
+                log::info!("Call {}", i + 1);
+            }
+
+            call().context("Call failed")?;
+        }
+
+        Ok(())
+    }
+
+    fn load_runtime(
+        &self,
+        rune: &[u8],
+        img: BaseImage,
+    ) -> Result<Box<dyn FnMut() -> Result<(), Error>>, Error> {
         if self.wasm3 {
             let mut runtime =
                 hotg_rune_wasm3_runtime::Runtime::load(&rune, img)
                     .context("Unable to initialize the virtual machine")?;
 
-            for i in 0..self.repeats {
-                if i > 0 {
-                    log::info!("Call {}", i + 1);
-                }
-                runtime.call().context("Call failed")?;
-            }
+            Ok(Box::new(move || runtime.call()))
         } else {
             let mut runtime =
                 hotg_rune_wasmer_runtime::Runtime::load(&rune, img)
                     .context("Unable to initialize the virtual machine")?;
 
-            for i in 0..self.repeats {
-                if i > 0 {
-                    log::info!("Call {}", i + 1);
-                }
-                runtime.call().context("Call failed")?;
-            }
+            Ok(Box::new(move || runtime.call()))
         }
-
-        Ok(())
     }
 
     fn initialize_image(&self, rune: &[u8]) -> Result<BaseImage, Error> {